add run left/right add toutes les rooms change rooms démo
